Issue #17168: Moved IT Module Inputs to Compilable Folder
