--- src/it/resources-noncompilable/org/checkstyle/suppressionxpathfilter/recordtypeparametername/InputXpathRecordTypeParameterNameTypeDefault.java
@@ -1,6 +1,0 @@
-//non-compiled with javac: Compilable with Java17
-package com.puppycrawl.tools.checkstyle.checks.naming.classtypeparametername;
-
-public record InputXpathRecordTypeParameterNameTypeDefault<t>(Integer x, String str) { // warn
-
-}
