+++ src/it/resources/org/checkstyle/suppressionxpathfilter/recordcomponentname/InputXpathRecordComponentNameDefault.java
@@ -1,0 +1,8 @@
+// Java17
+package org.checkstyle.suppressionxpathfilter.recordcomponentname;
+
+/* Config: default
+ */
+public record InputXpathRecordComponentNameDefault(int _value) { // warn
+
+}
