+++ src/it/resources/org/checkstyle/suppressionxpathfilter/recordtypeparametername/InputXpathRecordTypeParameterNameTypeDeclared.java
@@ -1,8 +1,11 @@
-//non-compiled with javac: Compilable with Java17
-package com.puppycrawl.tools.checkstyle.checks.naming.classtypeparametername;
+// Java17
+package org.checkstyle.suppressionxpathfilter.recordtypeparametername;
 
 import java.io.Serializable;
 import java.util.LinkedHashMap;
+
+import org.junit.platform.engine.support.hierarchical.Node;
+
 
 record InputXpathRecordTypeParameterNameTypeDeclared <foo extends Serializable & Cloneable> // warn
 (LinkedHashMap<String, Node> linkedHashMap) {
