+++ src/it/resources/org/checkstyle/suppressionxpathfilter/recordtypeparametername/InputXpathRecordTypeParameterNameTypeDefault.java
@@ -1,0 +1,6 @@
+// Java17
+package org.checkstyle.suppressionxpathfilter.recordtypeparametername;
+
+public record InputXpathRecordTypeParameterNameTypeDefault<t>(Integer x, String str) { // warn
+
+}
