+++ src/it/resources/org/checkstyle/suppressionxpathfilter/sealedshouldhavepermitslist/InputXpathSealedShouldHavePermitsListTopLevel.java
@@ -1,4 +1,4 @@
-//non-compiled with javac: Compilable with Java17
+// Java17
 package org.checkstyle.suppressionxpathfilter.sealedshouldhavepermitslist;
 
 public sealed class InputXpathSealedShouldHavePermitsListTopLevel { // warn
